fix: fix cache serialization type mismatch for daily word
